[Refactor]: 오거니즘 파일 생성, 구조화
