--- src/components/template/MessageSelectTemplate.jsx
@@ -1,11 +1,0 @@
-import MessagePaperSelect from "../organisms/MessagePaperSelect";
-
-const MessagePaperSelectTemplate = () => {
-  return (
-    <>
-      <MessagePaperSelect />
-    </>
-  );
-};
-
-export default MessagePaperSelectTemplate;
