+++ src/components/template/MessageWriteTemplate.jsx
@@ -1,11 +1,7 @@
-import React, { useEffect, useState } from "react";
-import { useSelector } from "react-redux";
-import MessagePaperSelect from "../organisms/MessagePaperSelect";
-import Message from "../common/Message";
-import TodayQuestion from "../common/TodayQuestion";
-import { BASE_URL } from "../../utils/URL";
-import axios from "axios";
-import sendIcon from "../../assets/sendIcon.svg";
+import React from "react";
+import TodayQuestion from "../organisms/Message/TodayQuestionOrganism";
+import MessagePaperSelect from "../organisms/Message/MessagePaperThumbnailOrganism";
+import MessageWriteOrganism from "../organisms/Message/MessageWriteOrganism";
 import styled from "styled-components";
 
 const ContentLayout = styled.div`
@@ -22,66 +18,14 @@ const ContentLayout = styled.div`
   }
 `;
 
-const CheckboxAndSendButton = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
 const MessageWriteTemplate = () => {
-  const messageInput = useSelector((state) => state.messageInput);
-  const selectedPaperNum = useSelector((state) => state.selectedPaperNum);
-  const [isChecked, setIsChecked] = useState(false);
-  const [isSent, setIsSent] = useState(false);
-
-  const handleCheckAnonymous = () => {
-    setIsChecked(!isChecked);
-  };
-
-  const handleSendButton = () => {
-    setIsSent(true);
-  };
-
-  let paramsToSend = {
-    context: messageInput,
-    int: selectedPaperNum,
-    // Header Authorization
-    anonymous: isChecked,
-  };
-
-  useEffect(() => {
-    axios
-      .post(`${BASE_URL}/message/create`, paramsToSend)
-      .then((response) => {
-        console.log(`Your message has been sent successfuly.`);
-      })
-      .catch((error) => {
-        console.error(`An error occured while sending the message.`);
-      });
-  }, [isSent]);
-
   return (
     <>
+      <div>프로필과 상태메시지</div>
       <ContentLayout>
         <TodayQuestion />
         <MessagePaperSelect />
-        <div>
-          <Message />
-          <CheckboxAndSendButton>
-            <form>
-              <label>익명으로 보내기 </label>
-              <input
-                type="checkbox"
-                checked={isChecked}
-                onChange={handleCheckAnonymous}
-              />
-            </form>
-            <button onClick={handleSendButton}>
-              메시지 보내기
-              <img src={sendIcon} alt="sendIcon" />
-            </button>
-          </CheckboxAndSendButton>
-        </div>
+        <MessageWriteOrganism />
       </ContentLayout>
     </>
   );
